Refactor persistence in a separate module
